backend/app/api.py
from fastapi import APIRouter, WebSocket, WebSocketDisconnect, HTTPException
from pydantic import BaseModel

# Import agent logic if needed
# from .agent import create_task_list, execute_tasks, review_and_repair, final_review

router = APIRouter()

# Example Pydantic model for chat input validation [cite: 170]
class ChatInput(BaseModel):
    query: str

# Example HTTP POST endpoint (alternative or complementary to WebSocket)
@router.post("/chat")
async def chat_endpoint(chat_input: ChatInput):
    try:
        # Placeholder: Initiate agent workflow here
        # task_file = await create_task_list(chat_input.query)
        # await execute_tasks(task_file) # Needs refinement for async/await results
        # final_response = await final_review(task_file)
        final_response = f"Received and processed (HTTP): {chat_input.query}" # Placeholder
        return {"response": final_response}
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

# WebSocket logic can also reside here or be called from main.py's WebSocket endpoint

# Example WebSocket handler function (could be called from main.py)
async def handle_chat_session(user_query: str, websocket: WebSocket):
    """
    Manages the agent workflow for a given user query via WebSocket.
    """
    try:
        await websocket.send_text("Agent: Creating task list...")
        # task_file_path = await create_task_list(user_query)
        # await websocket.send_text(f"Agent: Task list created at {task_file_path}")

        await websocket.send_text("Agent: Starting task execution (placeholder)...")
        # This part needs careful design for iterative execution, tool calls,
        # review/repair loop, and sending intermediate updates via WebSocket.
        # Example update:
        # await websocket.send_text("Agent: Executing task 1...")
        # tool_output = await call_tool(...)
        # await websocket.send_text(f"Agent: Task 1 output: {tool_output}")
        # await review_and_repair(...)
        # await websocket.send_text("Agent: Task 1 reviewed.")

        await websocket.send_text("Agent: Performing final review (placeholder)...")
        # final_response = await final_review(task_file_path)
        final_response = f"This is the final response for: {user_query}" # Placeholder
        await websocket.send_text(f"Agent: Final Response: {final_response}")

    except Exception as e:
        error_message = f"An error occurred: {e}"
        print(error_message)
        await websocket.send_text(f"Agent Error: {error_message}")

# Note: The actual implementation of the agent loop (create_task_list, execute_tasks, etc.)
# would reside in agent.py and be called from here or main.py.
# Real-time updates require the agent functions to 'yield' or send updates
# back to the WebSocket handler (e.g., using asyncio Queues or callbacks).